Migrated EtoileFoundation to CGFloat
git-svn-id: http://svn.gna.org/svn/etoile/trunk/Etoile/Frameworks/EtoileFoundation@8553 7b3f36aa-e6db-49c1-8b4c-66eac1997e64

+++ Source/ETByteSizeFormatter.m
@@ -32,7 +32,7 @@ MB, two digits for GB and three digits for TB. */
 
 	NSArray *sizeUnits = A(_(@"B"), _(@"KB"), _(@"MB"), _(@"GB"), _(@"TB"));
 	int nbOfUnits = [sizeUnits count];
-	float value = [anObject floatValue];
+	CGFloat value = [anObject floatValue];
   	int unitLevel = 0;
 
 	while (value >= 1024 && unitLevel < nbOfUnits)
